Bug fix - add usernames/fixed logout (#103)
* add usernames within callback

* signout session fix

* removed auth component
