+++ reactapp/src/actions/index.js
@@ -6,5 +6,4 @@ export * from './searchActions';
 export * from './settingActions';
 export * from './sendgridActions';
 export * from './loggedInActions';
-export * from './usernameActions';
