+++ server/routes/authRoutes.js
@@ -14,31 +14,48 @@ router.get(
 	passport.authenticate('auth0', {
 		scope: 'openid email profile'
 	}),
-	function (req, res) {
+	function(req, res) {
 		res.redirect('/');
 	}
 );
 
-router.get('/callback', function (req, res, next) {
-	passport.authenticate('auth0', function (err, user, info) {
+router.get('/callback', function(req, res, next) {
+	passport.authenticate('auth0', function(err, user, info) {
 		if (err) {
 			return next(err);
 		}
 		if (!user) {
 			return res.redirect('/signin');
 		}
-		req.logIn(user, function (err) {
+		req.logIn(user, function(err) {
 			if (err) {
 				return next(err);
 			}
 			const returnTo = req.session.returnTo;
 			delete req.session.returnTo;
-			const role =
-				req.user.profile._json['https://ratemydiy.herokuapp.com/roles'];
+			const sub = req.user.profile._json.sub;
+			const auth_id = sub.split('|')[1];
+			const username = req.user.profile._json.nickname;
+			const img_url = req.user.profile._json.picture;
+			const user = {
+				auth_id,
+				username,
+				img_url
+			};
+			const role = req.user.profile._json['https://ratemydiy.herokuapp.com/roles'];
 			if (role[0] === 'new') {
-				res.redirect(
-					(process.env.FRONTEND_URL || `http://localhost:3000`) + `/signin`
-				);
+				usersDB
+					.addUser(user)
+					.then(dbRes => {
+						console.log('DB SUCCESS');
+						res.redirect(
+							returnTo || process.env.FRONTEND_URL || `http://localhost:3000`
+						);
+					})
+					.catch(dbErr => {
+						console.log('DB ERROR');
+						res.status(500).json({ error: 'Could not add user to database' });
+					});
 			} else {
 				res.redirect(
 					returnTo || process.env.FRONTEND_URL || `http://localhost:3000`
@@ -48,7 +65,7 @@ router.get('/callback', function (req, res, next) {
 	})(req, res, next);
 });
 
-router.get('/loggedIn', function (req, res, next) {
+router.get('/loggedIn', function(req, res, next) {
 	// console.log('cookies:', req.cookies);
 	// console.log('user:', req.user);
 
@@ -71,10 +88,15 @@ router.get('/loggedIn', function (req, res, next) {
 
 router.get('/signout', (req, res) => {
 	req.logout();
-	res.redirect(process.env.FRONTEND_URL || `http://localhost:3000`);
+	req.session.destroy();
+	res.redirect(
+		`https://ratemydiy.auth0.com/v2/logout?returnTo=${process.env
+			.FRONTEND_URL ||
+			`http://localhost:3000`}&client_id=ivcQqf9pMrudlWLh8zvTihsPN04bzBJs`
+	);
 });
 
-router.post('/test', ensureLoggedIn, function (req, res, next) {
+router.post('/test', ensureLoggedIn, function(req, res, next) {
 	console.log('cookies:', req.cookies);
 	console.log('user:', req.user);
 
@@ -83,7 +105,7 @@ router.post('/test', ensureLoggedIn, function (req, res, next) {
 	res.status(200).json({ message: 'it works' });
 });
 
-router.get('/cookie', function (req, res, next) {
+router.get('/cookie', function(req, res, next) {
 	console.log(req.cookies);
 	res.status(200).json(req.cookies);
 });
